+++ lib/blocks.php
@@ -76,7 +76,7 @@ function gutenberg_reregister_core_block_types() {
 					'post-excerpt.php'              => 'core/post-excerpt',
 					'post-featured-image.php'       => 'core/post-featured-image',
 					'post-hierarchical-terms.php'   => 'core/post-hierarchical-terms',
-					'post-navigation.php'           => 'core/post-navigation',
+					'post-navigation-link.php'      => 'core/post-navigation-link',
 					'post-tags.php'                 => 'core/post-tags',
 					'post-title.php'                => 'core/post-title',
 					'query.php'                     => 'core/query',
