+++ packages/block-library/src/post-navigation-link/block.json
@@ -1,6 +1,6 @@
 {
 	"apiVersion": 2,
-	"name": "core/post-navigation",
+	"name": "core/post-navigation-link",
 	"category": "design",
 	"attributes": {
 		"type": {
