+++ packages/block-library/src/post-navigation-link/variations.js
@@ -2,34 +2,37 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { postNavigationNext, postNavigationPrevious } from '@wordpress/icons';
+import {
+	postNavigationLinkNext,
+	postNavigationLinkPrevious,
+} from '@wordpress/icons';
 
 const variations = [
 	{
 		isDefault: true,
-		name: 'post-navigation-next',
+		name: 'post-navigation-link-next',
 		title: __( 'Next post link' ),
 		description: __(
 			'Displays the next post link that is adjacent to the current post.'
 		),
-		icon: postNavigationNext,
+		icon: postNavigationLinkNext,
 		attributes: { type: 'next' },
 		scope: [ 'inserter', 'transform' ],
 	},
 	{
-		name: 'post-navigation-previous',
+		name: 'post-navigation-link-previous',
 		title: __( 'Previous post link' ),
 		description: __(
 			'Displays the previous post link that is adjacent to the current post.'
 		),
-		icon: postNavigationPrevious,
+		icon: postNavigationLinkPrevious,
 		attributes: { type: 'previous' },
 		scope: [ 'inserter', 'transform' ],
 	},
 ];
 
 /**
- * Add `isActive` function to all `post-navigation` variations, if not defined.
+ * Add `isActive` function to all `post-navigation-link` variations, if not defined.
  * `isActive` function is used to find a variation match from a created
  *  Block by providing its attributes.
  */
