+++ packages/e2e-tests/fixtures/blocks/core__post-navigation-link.html
@@ -1,0 +1,1 @@
+<!-- wp:post-navigation-link {"type":"next"} /-->
