+++ packages/e2e-tests/fixtures/blocks/core__post-navigation-link.json
@@ -1,7 +1,7 @@
 [
     {
         "clientId": "_clientId_0",
-        "name": "core/post-navigation",
+        "name": "core/post-navigation-link",
         "isValid": true,
         "attributes": {
             "type": "next",
