+++ packages/e2e-tests/fixtures/blocks/core__post-navigation-link.parsed.json
@@ -1,6 +1,6 @@
 [
     {
-        "blockName": "core/post-navigation",
+        "blockName": "core/post-navigation-link",
         "attrs": {
             "type": "next"
         },
