+++ packages/e2e-tests/fixtures/blocks/core__post-navigation-link.serialized.html
@@ -1,0 +1,1 @@
+<!-- wp:post-navigation-link /-->
