--- packages/e2e-tests/fixtures/blocks/core__post-navigation.html
@@ -1,1 +1,0 @@
-<!-- wp:post-navigation {"type":"next"} /-->
