--- packages/e2e-tests/fixtures/blocks/core__post-navigation.serialized.html
@@ -1,1 +1,0 @@
-<!-- wp:post-navigation /-->
