+++ packages/icons/src/index.js
@@ -141,8 +141,8 @@ export { default as postExcerpt } from './library/post-excerpt';
 export { default as postFeaturedImage } from './library/post-featured-image';
 export { default as postList } from './library/post-list';
 export { default as postTitle } from './library/post-title';
-export { default as postNavigationPrevious } from './library/post-navigation-previous';
-export { default as postNavigationNext } from './library/post-navigation-next';
+export { default as postNavigationLinkPrevious } from './library/post-navigation-link-previous';
+export { default as postNavigationLinkNext } from './library/post-navigation-link-next';
 export { default as preformatted } from './library/preformatted';
 export { default as pullLeft } from './library/pull-left';
 export { default as pullRight } from './library/pull-right';
